feat: remove covered tests [DHIS2-16799]
