+++ dhis-2/dhis-test-web-api/src/test/java/org/hisp/dhis/webapi/controller/userdatastore/UserDatastoreFilterControllerTest.java
@@ -33,6 +33,7 @@ import java.util.List;
 import javax.servlet.http.HttpServletResponse;
 import org.hisp.dhis.datastore.DatastoreParams;
 import org.hisp.dhis.web.HttpStatus;
+import org.hisp.dhis.webapi.controller.AbstractDataValueControllerTest;
 import org.hisp.dhis.webapi.controller.UserDatastoreController;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -46,7 +47,8 @@ import org.junit.jupiter.api.Test;
  *
  * @author Jan Bernitt
  */
-class UserDatastoreFilterControllerTest extends AbstractUserDatastoreControllerTest {
+class UserDatastoreFilterControllerTest
+    extends AbstractDataValueControllerTest.AbstractUserDatastoreControllerTest {
   @BeforeEach
   void setUp() {
     // simple values
